requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django_rest_passwordreset==1.3.0
pyyaml==6.0
requests==2.28.2
ujson==5.7.0
celery[redis]==5.2.7
redis==4.5.4